bot/requirements.txt
aiocache==0.12.3
annotated-types==0.7.0
betterlogging==1.0.0
cffi==1.17.1
cloudpickle==3.1.0
cryptography==44.0.0
gitdb==4.0.12
GitPython==3.1.44
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
psutil==6.1.0
pyaes==1.6.1
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
Pyrogram==2.0.106
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
rich==13.9.4
smmap==5.0.2
TgCrypto==1.2.5
typing_extensions==4.12.2
